requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
Flask==1.1.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.1
mendeleev==0.6.0
monty==3.0.2
mpmath==1.1.0
networkx==2.4
numpy==1.18.4
palettable==3.3.0
pandas==1.0.3
plotly==4.7.1
pyfiglet==0.8.post1
pymatgen==2020.4.29
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
qmpy-rester==0.2.0
requests==2.23.0
retrying==1.3.3
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scipy==1.4.1
six==1.14.0
spglib==1.15.1
SQLAlchemy==1.3.17
sympy==1.5.1
tabulate==0.8.7
urllib3==1.25.9
Werkzeug==1.0.1
